Update telegram brigde
- remove Group dependency to be able to run unit tests
- add Group handler function to service and pass Group from index->messages->bridge command -> service
- move Group model to correct models folder
- move bridge prefix command to correct commands folder
